requirements.txt
requests==2.32.5
Pillow>=10.0.0
python-dotenv==1.2.1
